Tons of edits to achieve feature parity with Obsidian.  More to come.

+++ apply.sh
@@ -20,10 +20,9 @@
 set -o nounset                              # Treat unset variables as an error
 
 zipfile=$(ls -1 *.zip 2> /dev/null)
-[[ ! -f "$zipfile" ]] && { echo "Can't find a patch zip file."; exit 1; }
+# [[ ! -f "$zipfile" ]] && { echo "Can't find a patch zip file."; exit 1; }
 echo "$zipfile"
-mv "$zipfile" patches/
-unzip "patches/$zipfile"
+[[ -f "$zipfile" ]] && { mv "$zipfile" patches/; unzip "patches/$zipfile"; }
 patchfile=$(ls -1 *.patch)
 echo "$patchfile"
 patch -p1 < "$patchfile"
